Moves HOC into utils folder
